requirements.txt
cached-property==1.3.0
frozendict==1.2
genanki==0.4.0
lxml==3.8.0
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pystache==0.5.4
python-pptx==0.6.6
PyYAML==3.12
XlsxWriter==0.9.8
